Do not fail on chown
When using compose with multiple services, some services may contain
mounts that others don't. In this case, chown will fail with 'No such
file or directory' since chown is currently the combination of the
volumes from every service. The easy fix for now is to just never fail
on chown.